backend/requirements.txt
fastapi==0.87.0
python-dotenv==0.21.0 
pydantic==1.10.2
pymongo==4.3.3
py2neo==2021.2.3
neo4j==5.2.1
motor==3.1.1
uvicorn==0.20.0
